+++ src/control/controllers/controllers/node_point_fitting.py
@@ -1,4 +1,4 @@
-from math import atan, atan2, degrees, floor, pi, sqrt
+from math import atan, atan2, degrees, pi, sqrt
 import time
 
 import cv2
@@ -14,7 +14,7 @@ from driverless_msgs.msg import Cone, ConeDetectionStamped
 from sensor_msgs.msg import Image
 
 from driverless_common.draw import *
-from driverless_common.point import Point
+from driverless_common.point import Point, cone_to_point, dist
 from driverless_common.shutdown_node import ShutdownNode
 
 from typing import List, Tuple
@@ -31,99 +31,45 @@ RIGHT_CONE_COLOUR = Cone.YELLOW
 
 WEIGHT = -1
 
-POINTS_PER_LINE = 10
-SPLINES = 11
-if SPLINES % 2 == 0:
-    SPLINES += 1
+
+def closest_point_on_curve(point: Point, curve: np.ndarray) -> Point:
+    """
+    Find the closest point on the curve to the cone
+    * param point: the cone coord
+    * param curve: the curve
+    * return: the closest point on the curve to the cone
+    """
+    closest_point = ORIGIN
+    min_dist = float("inf")
+    for i in range(len(curve)):
+        p = Point(i / 10, curve[i])
+        d = dist(point, p)
+        if d < min_dist:
+            min_dist = d
+            closest_point = p
+    return closest_point
 
 
-class PoinntFitController(Node):
+class BetterReactiveController(Node):
     prev_steering_angle: float = 0
     Kp_angle: float = 20
     targ_vel: float = 4  # m/s
     debug_img = np.zeros((HEIGHT, WIDTH, 3), dtype=np.uint8)  # create black image
 
-    steeringPaths: List[List[Point]] = []
-    line_len: float = 15.0  # len of spline line
-
     def __init__(self):
-        super().__init__("point_fit_controller_node")
+        super().__init__("reactive_traj_controller_node")
 
         # debug image
-        self.create_subscription(Image, "/debug_imgs/lidar_det_img", self.img_callback, 1)
+        self.create_subscription(Image, "/debug_imgs/vision_det_img", self.img_callback, 1)
         # cone detections
-        self.create_subscription(ConeDetectionStamped, "/lidar/cone_detection", self.callback, 1)
+        self.create_subscription(ConeDetectionStamped, "/slam/local_map", self.callback, 1)
 
         # publishers
         self.control_publisher: Publisher = self.create_publisher(AckermannDriveStamped, "/control/driving_command", 1)
 
         self.debug_img_publisher: Publisher = self.create_publisher(Image, "/debug_imgs/control_img", 1)
 
-        self.initialize_splines()  # initialize steering curves to take
-
         self.get_logger().info("---Better Reactive Controller Node Initalised---")
-
-    def initialize_splines(self):
-        # cicle will pass through 0,0 and and additional point:
-        # for each end point, determine the circle defining it with 10 points
-        end_points = []
-        end_points.append(Point(self.line_len / 2, -5 * self.line_len / 10))  # origin
-        end_points.append(Point(self.line_len / 2 + self.line_len / 10, -4 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + 2 * self.line_len / 10, -3 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + 3 * self.line_len / 10, -2 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + 4 * self.line_len / 10, -1 * self.line_len / 10))
-        end_points.append(Point(self.line_len, 0.0))
-        end_points.append(Point(self.line_len / 2 + 4 * self.line_len / 10, 1 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + 3 * self.line_len / 10, 2 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + 2 * self.line_len / 10, 3 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2 + self.line_len / 10, 4 * self.line_len / 10))
-        end_points.append(Point(self.line_len / 2, 5 * self.line_len / 10))
-
-        self.steering_paths = [[] for i in range(SPLINES)]  # create empty list of lists
-        for index in range(SPLINES):
-            self.steering_paths[index] = []
-
-            # straight up
-            if index == int(SPLINES / 2):
-                x = 0
-                y = 0
-                self.steering_paths[int(SPLINES / 2)] = []
-                for i in range(POINTS_PER_LINE):
-                    self.steering_paths[int(SPLINES / 2)].append(Point(x, y))  # store
-                    x = x + self.line_len / 9
-                continue
-
-            h = (end_points[index].x ** 2 + end_points[index].y ** 2) / (2 * end_points[index].y)  # center
-            r = abs(h)  # radius
-            circum = pi * 2 * r  # circumference
-            fraction = self.line_len / circum  # percentage of circle
-            angle = 2 * pi * fraction / 10.0  # angle / 10 in radians
-
-            # if the circle is to the right of the origin, the angle needs to be negative
-            if index > int(SPLINES / 2):
-                angle = -angle
-                r = -r
-
-            center = Point(0, h)  # center of circle
-            vector = Point(0, r)  # vector to point
-            print(
-                "angle: "
-                + str(angle)
-                + " r: "
-                + str(r)
-                + " h: "
-                + str(h)
-                + " center: "
-                + str(center)
-                + " vector: "
-                + str(vector)
-            )
-            for i in range(POINTS_PER_LINE):
-                self.steering_paths[index].append(Point(-(vector.x + center.x), vector.y + center.y))  # store
-                print("vector: " + str(vector))
-                vector = Point(
-                    cos(angle) * vector.x - sin(angle) * vector.y, sin(angle) * vector.x + cos(angle) * vector.y
-                )  # rotate
 
     def img_callback(self, img_msg: Image):
         self.debug_img = cv_bridge.imgmsg_to_cv2(img_msg, desired_encoding="bgr8")
@@ -136,57 +82,64 @@ class PoinntFitController(Node):
         speed = 0.0
         steering_angle = 0.0
 
-        cones: List[Cone] = [
-            c for c in cone_msg.cones if abs(c.location.y) < 7.0 and c.location.x < 12
-        ]  # only cones in front of car
+        cones: List[Cone] = cone_msg.cones
+        left_cones = [c for c in cones if c.color == LEFT_CONE_COLOUR]
+        right_cones = [c for c in cones if c.color == RIGHT_CONE_COLOUR]
 
-        best_spline = 0
-        if len(cones) != 0:
-            s_error = [0] * len(self.steering_paths)  # error for each spline
-            max_error = 1e20
+        if len(left_cones) != 0 and len(right_cones) != 0:
+            # compute 10 quadratic curves based on different steering angles
+            # find the best curve that reduces the error from curve to each cone
+            # use the steering angle of that curve
+            A: np.ndarray = np.linspace(-0.1, 0.1, 15)
+            x: np.ndarray = np.linspace(0, 10, 100)
+            curves: List[np.ndarray] = []
+            # compute the error for each curve
+            errors: List[float] = []
+            for a in A:
+                y = a * x**2
+                curves.append(y)
 
-            for n_spline, steering_path in enumerate(self.steering_paths):
-                n_spline_points = len(steering_path)
-                for cone in cones:
-                    # distance from origin
-                    distance = sqrt(cone.location.x**2 + cone.location.y**2)
+                error = 0
+                for cone in left_cones:
+                    # find the closest point on the curve to the cone
+                    closest_point: Point = closest_point_on_curve(cone_to_point(cone), y)
+                    # compute the error
+                    error += dist(cone_to_point(cone), closest_point) * WEIGHT / len(left_cones)
 
-                    # bin
-                    bin = min(n_spline_points - 1, int(round(distance / self.line_len * n_spline_points, 0)) - 1)
-                    if bin < 0:
-                        bin = 0
+                for cone in right_cones:
+                    # find the closest point on the curve to the cone
+                    closest_point: Point = closest_point_on_curve(cone_to_point(cone), y)
+                    # compute the error
+                    error += dist(cone_to_point(cone), closest_point) * -WEIGHT / len(right_cones)
 
-                    s_error[n_spline] += sqrt(
-                        (cone.location.x - steering_path[bin].x) ** 2 + (cone.location.y - steering_path[bin].y) ** 2
-                    )
+                errors.append(error)
 
-                if s_error[n_spline] < max_error:
-                    max_error = s_error[n_spline]
-                    best_spline = n_spline
-
-                for point in steering_path:
+                for i in range(len(x)):
                     # draw each element in target spline
                     cv2.drawMarker(
                         self.debug_img,
-                        loc_to_img_pt(point.x, point.y).to_tuple(),
+                        loc_to_img_pt(x[i], y[i]).to_tuple(),
                         (0, 0, 255),
                         markerType=cv2.MARKER_SQUARE,
-                        markerSize=2,
+                        markerSize=1,
                         thickness=2,
                     )
 
-            best_path = self.steering_paths[best_spline]
-            for point in best_path:
+            # find the best curve
+            best_curve: np.ndarray = curves[np.abs(errors).argmin()]
+            for i in range(len(x)):
                 # draw each element in target spline
                 cv2.drawMarker(
                     self.debug_img,
-                    loc_to_img_pt(point.x, point.y).to_tuple(),
+                    loc_to_img_pt(x[i], best_curve[i]).to_tuple(),
                     (0, 255, 0),
                     markerType=cv2.MARKER_SQUARE,
-                    markerSize=2,
+                    markerSize=1,
                     thickness=2,
                 )
             self.debug_img_publisher.publish(cv_bridge.cv2_to_imgmsg(self.debug_img, encoding="bgr8"))
+
+            steering_angle: float = A[np.abs(errors).argmin()] * self.Kp_angle
 
             speed = self.targ_vel
 
@@ -202,7 +155,7 @@ class PoinntFitController(Node):
 
 def main(args=None):
     rclpy.init(args=args)
-    node = PoinntFitController()
+    node = BetterReactiveController()
     rclpy.spin(node)
     node.destroy_node()
     rclpy.shutdown()
