+++ src/control/controllers/controllers/node_vector_spline.py
